requirements.txt
aerich==0.9.1
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
asyncclick==8.2.2.2
asyncpg==0.30.0
atlastk==0.13.5
click==8.3.0
dictdiffer==0.9.0
fastapi==0.117.1
h11==0.16.0
idna==3.10
iso8601==2.1.0
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
pyotp==2.9.0
pypika-tortoise==0.6.2
python-dotenv==1.1.1
pytz==2025.2
sniffio==1.3.1
starlette==0.48.0
tortoise==0.1.1
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
